Serve index.html implicity if directory requests fail
This is a hack for pip.  Tools like s3pypi build index.html files that
pip can use to find out what versions of a package are available.  pip
expects the web server to serve an index page, though, if it makes a
request for a directory (like "/package-name/") so we need to make
another request to see if there is an index.html page and serve it if
there is one.